+++ packages/side-effects/side-effect-aria-live/types/index.d.ts
@@ -1,0 +1,3 @@
+import { Observer, Emitted } from "@curi/router";
+export declare type AriaLiveValue = "assertive" | "polite" | "off";
+export default function createAriaLiveSideEffect(fmt: (emitted: Emitted) => string, mode?: AriaLiveValue): Observer;