+++ website/src/client/pages/Packages/index.js
@@ -2,6 +2,7 @@ import Router from "./Router";
 import RouteActive from "./RouteActive";
 import RouteAncestors from "./RouteAncestors";
 import RoutePrefetch from "./RoutePrefetch";
+import SideEffectAriaLive from "./SideEffectAriaLive";
 import SideEffectTitle from "./SideEffectTitle";
 import SideEffectScroll from "./SideEffectScroll";
 import ReactPkg from "./ReactPkg";
@@ -14,6 +15,7 @@ export default {
   "route-active": RouteActive,
   "route-ancestors": RouteAncestors,
   "route-prefetch": RoutePrefetch,
+  "side-effect-aria-live": SideEffectAriaLive,
   "side-effect-title": SideEffectTitle,
   "side-effect-scroll": SideEffectScroll,
   react: ReactPkg,